nvidia-gpu: fix log typos and minor cleanups
Address miscellaneous CodeRabbit review feedback in a single
batch: log message typos, defensive checks, and minor type
tightening. Each fix is independent and small; bundling avoids
inflating the per-finding patch count for purely cosmetic
changes.

- NvidiaPcieDevice: drop duplicated "for for" in error message
- NvidiaGpuPowerSensor: "Temperature Sensor" -> "Power Sensor"
- NvidiaDeviceDiscovery: nicNetworkPortCount uint64_t -> uint16_t,
  and matching loadVariant<> instantiation and loop counter
- NvidiaDriverInformation: silence discarded set_property bool
  result via (void) cast
- NvidiaGpuCurrentUtilization: tighten size check from < 8 to
  < sizeof(uint32_t) to match what is actually unpacked
- NvidiaLongRunningHandler: mark decoded but unused instanceId
  [[maybe_unused]]

Tested:
Pure refactor / cleanup with no observable runtime impact;
observed behavior is the same as before. Correctness of the
fix is verified by inspection.

+++ src/nvidia-gpu/NvidiaDeviceDiscovery.cpp
@@ -405,7 +405,7 @@ void processSensorConfigs(
             try
             {
                 configs.nicNetworkPortCount =
-                    loadVariant<uint64_t>(cfg, "NicNetworkPortCount");
+                    loadVariant<uint16_t>(cfg, "NicNetworkPortCount");
             }
             catch (const std::invalid_argument&)
             {
