+++ src/nvidia-gpu/NvidiaDeviceDiscovery.hpp
@@ -24,7 +24,7 @@ struct SensorConfigs
 {
     std::string name;
     uint64_t pollRate{};
-    uint64_t nicNetworkPortCount{};
+    uint16_t nicNetworkPortCount{};
 };
 
 class GpuDevice;
